Add files via upload
MultiStageEncryption is a java application which executes AES and 3DES application for any type of files. The files are first encoded to convert it into a text file and then encryption is done (AES followed by 3DES).
Decryption is performed in a reverse order 3DES decryption first and then AES decryption before decoding the file.